Wood, Parrot, Bird
Made the prefab for the bird, the sign. Added a method to  set the wood inactive after interacting with the bird via the mouse and after colliding with it. Made the signs print a message, but still need dialog options from the bird and the signs. Will add a particle system to occur when the wood is collected.